requirements.txt
bcrypt==4.0.1
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.5.3.56
pygame==2.3.0
RPi.GPIO==0.7.1
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
